Add unit parse (#324)
* move UnitParse functions to their own file

* remove misc.cpp
