+++ model/CMakeLists.txt
@@ -14,6 +14,7 @@ configure_file(
 
 # Don't use aux_source_directory on . because we don't want to compile openMalaria.cpp in to the lib.
 set (Model_CPP
+  sim.cpp
   # Simulator.cpp
   Population.cpp
   PopulationAgeStructure.cpp
@@ -87,8 +88,8 @@ set (Model_CPP
   util/sampler.cpp
   util/SpeciesIndexChecker.cpp
   util/DocumentLoader.cpp
-  util/misc.cpp
   util/random.cpp
+  util/UnitParse.cpp
   
   interventions/InterventionManager.cpp
   interventions/ITN.cpp
