+++ model/mon/Continuous.cpp
@@ -22,7 +22,7 @@
 #include "mon/info.h"   // lineEnd
 #include "util/errors.h"
 #include "util/CommandLine.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 #include "schema/monitoring.h"
 
 #include <vector>
