+++ model/util/UnitParse.cpp
@@ -22,7 +22,7 @@
 #include "util/errors.h"
 #include "util/ModelOptions.h"
 #include "util/CommandLine.h"
-#include "util/timeConversions.h"
+#include "util/UnitParse.h"
 #include "schema/scenario.h"
 #include "mon/management.h"
 
@@ -32,40 +32,8 @@
 
 namespace OM {
 
-// ———  SimTime stuff  ———
-
-// Scenario constants
-int SimData::interval;
-size_t SimData::steps_per_year;
-double SimData::years_per_step;
-
-SimDate sim::s_start;
-SimDate sim::s_end;
-
-SimTime sim::s_max_human_age;
-
-// Global variables
-#ifndef NDEBUG
-bool sim::in_update = false;
-#endif
-SimTime sim::s_t0;
-SimTime sim::s_t1;
-
-SimTime sim::s_interv;
-
-using util::CommandLine;
-
-
 // ———  Unit parsing stuff  ———
 namespace UnitParse {
-
-static int monthLen[] = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
-static int monthStart[] = { // offsets of start of month from start of year
-    0 /*Jan*/, 31, 59, 90,
-    120 /*May*/, 151, 181, 212,
-    243 /*Sept*/, 273, 304, 334,
-    365 /*next year: stop condition in formatDate*/
-};
 
 template<typename T>
 int castToInt( T x ){
@@ -89,7 +57,7 @@ SimTime readShortDuration( const std::string& str, DefaultUnit defUnit ){
         }else if( defUnit == NONE ){
             throw util::format_error( "unit required but not given (try e.g. 5d or 12t)" );
         }else{
-            if( CommandLine::option(CommandLine::DEPRECATION_WARNINGS) ){
+            if( util::CommandLine::option(util::CommandLine::DEPRECATION_WARNINGS) ){
                 cerr << "Deprecation warning: duration \"" << str << "\" specified without unit; it is recommended to do so (e.g. 5d or 1t)" << endl;
             }
             if( defUnit == DAYS ){
@@ -133,7 +101,7 @@ double parseDurationAndUnit( const std::string& str, DefaultUnit& unit ){
         if( v == 0.0 ) unit = DAYS;        // special case: 0 does not require a unit; pretend the unit is days
         else if( unit == NONE /*no default set for this value*/ ){
             throw util::format_error( "unit required but not given (try e.g. 5d or 12t or 2.3y)" );
-        }else if( CommandLine::option(CommandLine::DEPRECATION_WARNINGS) ){
+        }else if( util::CommandLine::option(util::CommandLine::DEPRECATION_WARNINGS) ){
             cerr << "Deprecation warning: duration \"" << str << "\" specified without unit; it is recommended to do so (e.g. 5d or 1t or 0.5y)" << endl;
         }
     }else{
@@ -199,7 +167,7 @@ SimDate readDate( const std::string& str, DefaultUnit defUnit ){
         }
         return date;
     }else{
-        if( CommandLine::option(CommandLine::DEPRECATION_WARNINGS) ){
+        if( util::CommandLine::option(util::CommandLine::DEPRECATION_WARNINGS) ){
             cerr << "Deprecation warning: time specified via duration \"" << str << "\" where a date could be used; recommended to use a date (e.g. 2011-12-20)" << endl;
         }
         return sim::startDate() + readDuration( str, defUnit );
@@ -207,64 +175,5 @@ SimDate readDate( const std::string& str, DefaultUnit defUnit ){
 }
 
 }       // end of UnitParse namespace
-
-
-ostream& operator<<( ostream& stream, SimTime time ){
-    if( time.inDays() % sim::DAYS_IN_YEAR == 0 ){
-        stream << time.inYears() << 'y';
-    } else {
-        stream << time.inDays() << 'd';
-    }
-    return stream;
-}
-
-ostream& operator<<( ostream& stream, SimDate date ){
-    int days = date.d;
-    if (days < 0) {
-        // Shouldn't happen; best still to print something
-        stream << days << 'd';
-    } else {
-        int year = days / sim::DAYS_IN_YEAR;
-        days -= year * sim::DAYS_IN_YEAR;
-        
-        int month = 0;
-        while( days >= UnitParse::monthStart[month+1] ) ++month;
-        days -= UnitParse::monthStart[month];
-        assert( month < 12 && days < UnitParse::monthLen[month] );
-        
-        // Inconsistency in year vs month and day: see parseDate()
-        stream << setfill('0') << setw(4) << year << '-'
-                << setw(2) << (month+1) << '-' << (days+1)
-                << setw(0) << setfill(' ');
-    }
-    return stream;
-}
-
-void sim::init( const scnXml::Scenario& scenario ){
-    SimData::interval = scenario.getModel().getParameters().getInterval();
-    SimData::steps_per_year = SimTime::oneYear().inSteps();
-    SimData::years_per_step = 1.0 / SimData::steps_per_year;
-    
-    sim::s_max_human_age =
-        SimTime::fromYearsD( scenario.getDemography().getMaximumAgeYrs() );
-    
-    sim::s_start = SimDate::origin();
-    auto mon = scenario.getMonitoring();
-    if( mon.getStartDate().present() ){
-        try{
-            // on failure, this throws or returns SimDate::never()
-            sim::s_start = UnitParse::parseDate( mon.getStartDate().get() );
-            if( sim::s_start == SimDate::never() ){
-                throw util::format_error( "invalid format (expected YYYY-MM-DD)" );
-            }
-        }catch( const util::format_error& e ){
-            throw util::xml_scenario_error( string("monitoring/startDate: ").append(e.message()) );
-        }
-    }
-    
-    sim::s_interv = SimTime::never();    // large negative number
-    
-    sim::s_end = mon::readSurveyDates( mon );
-}
 
 }
